requirements.txt
# Hepsiburada Test Automation Requirements
# Core testing framework
pytest>=7.0.0
pytest-html>=3.1.0

# Selenium WebDriver
selenium>=4.10.0
webdriver-manager>=3.8.0

# Test reporting
pytest-html>=3.1.0
allure-pytest>=2.12.0

# Test retry mechanism
pytest-rerunfailures>=12.0

# Screenshot capture on failure
pytest-screenshot-on-failure>=1.0.0

# Logging
loguru>=0.7.0

# Date and time handling
python-dateutil>=2.8.0

# Regular expressions
regex>=2023.0.0

# HTTP requests
requests>=2.31.0

# Configuration management
python-dotenv>=1.0.0